feat: separate logic from cart component into custom hook
