+++ src/constants/constants.ts
@@ -1,3 +1,7 @@
 export const dog1 = 'https://blogs.biomedcentral.com/on-biology/wp-content/uploads/sites/5/2018/04/nature-3042751_1920-copy-2-620x342.jpg'
 export const dog2 = 'https://cdn.pixabay.com/photo/2019/09/23/16/47/friendship-4499076__340.jpg'
 export const dog3 = 'https://www.akc.org/wp-content/uploads/2017/12/Yellow-Lab-High-Five.jpg'
+
+export const aboutText1 = "Here at the dog store, we're passionate about providing the best possible nutrition and supplies for dogs. As dog owners ourselves, we understand how important it is to ensure that our furry friends are healthy and happy. That's why we've made it our mission to offer premium quality dog food and supplies that are tailored to meet the unique needs of each individual dog."
+
+export const aboutText2 = "We believe that every dog deserves the very best, which is why we source only the highest quality ingredients for our dog food products. Our team of experts carefully selects each ingredient based on its nutritional value and suitability for dogs, ensuring that our food is not only delicious but also packed with essential nutrients."
